Global interlanguage links feature
To use, set the new global $wgInterwikiCentralInterlanguageDB to the
name of a MediaWiki DB from which to fetch global interlanguage links.
For example, on Uncyclomedia projects this could be 'uncm_meta' for
most of the Uncyclopedias to fetch their interlanguage links from the
meta-wiki, but perhaps 'illg_en' for the Illogicopedias to fetch their
interlanguage links from the English Illogicopedia's DB (instead of
the meta Uncyclomedia wiki's DB).

Disables local interlanguage links for as long as it is enabled, and
as such likewise prevents their editing, but does not change anything
currently in the local table.

Also bumped the version number and got rid of the silly, outdated
datestamp in it.

Bug: T220247
Change-Id: I6b691ef8e37367fe0e0fcb8b3778f4470d0c5200