Don't register InterwikiLoadPrefix hook unless it is needed
The InterwikiLoadPrefix hook has been deprecated, as it doesn't
support some functions needed in a round-trip wikitext parser
(T270444).  This extension doesn't require the hook in the way that it
is used in WMF production (with a static interwiki cache), so don't
register the hook in this configuration to avoid triggering the
deprecation warning.

Bug: T270444
Depends-On: If2507017c99c4ee42c104a0890bc45a84d7239d5
Change-Id: Id5c02a824c6d6a97200074430aa7d5c1db804396